feat: implemented horizontal scaling
