+++ nodemon.json
@@ -1,6 +1,5 @@
 {
   "watch": ["src"],
   "ext": ".ts",
-  "ignore": [],
-  "exec": "ts-node ./src/index.ts"
+  "ignore": []
 }
